+++ abjadext/nauert/__init__.py
@@ -1,21 +1,34 @@
 """
 Extension for quantizing rhythm, based on Paul Nauert's Q-Grid technique.
 """
-from .attackpointoptimizers import AttackPointOptimizer, MeasurewiseAttackPointOptimizer, NaiveAttackPointOptimizer, NullAttackPointOptimizer
-from .gracehandlers import CollapsingGraceHandler, ConcatenatingGraceHandler, DiscardingGraceHandler, GraceHandler
+from .attackpointoptimizers import (
+    AttackPointOptimizer,
+    MeasurewiseAttackPointOptimizer,
+    NaiveAttackPointOptimizer,
+    NullAttackPointOptimizer,
+)
+from .gracehandlers import (
+    CollapsingGraceHandler,
+    ConcatenatingGraceHandler,
+    DiscardingGraceHandler,
+    GraceHandler,
+)
 from .heuristics import DistanceHeuristic, Heuristic
-from .jobhandlers import JobHandler, ParallelJobHandler, ParallelJobHandlerWorker, SerialJobHandler
+from .jobhandlers import (
+    JobHandler,
+    ParallelJobHandler,
+    ParallelJobHandlerWorker,
+    SerialJobHandler,
+)
 from .qevents import PitchedQEvent, QEvent, SilentQEvent, TerminalQEvent
-from .QEventProxy import QEventProxy
-from .QEventSequence import QEventSequence
-from .QGrid import QGrid
-from .QGridContainer import QGridContainer
-from .QGridLeaf import QGridLeaf
+from .qeventproxy import QEventProxy
+from .qeventsequence import QEventSequence
+from .qgrid import QGrid, QGridContainer, QGridLeaf
 from .qschemas import BeatwiseQSchema, MeasurewiseQSchema, QSchema
 from .qschemaitems import BeatwiseQSchemaItem, MeasurewiseQSchemaItem, QSchemaItem
-from .qtargets import BeatwiseQTarget, MeasurewiseQTarget, QTarget
 from .QTargetBeat import QTargetBeat
 from .QTargetMeasure import QTargetMeasure
+from .qtargets import BeatwiseQTarget, MeasurewiseQTarget, QTarget
 from .QuantizationJob import QuantizationJob
 from .Quantizer import Quantizer
 from .searchtrees import SearchTree, UnweightedSearchTree, WeightedSearchTree
