+++ abjadext/nauert/qtargets.py
@@ -5,7 +5,6 @@ import typing
 
 import abjad
 
-from .QEventSequence import QEventSequence
 from .QTargetBeat import QTargetBeat
 from .QTargetMeasure import QTargetMeasure
 from .attackpointoptimizers import AttackPointOptimizer, NaiveAttackPointOptimizer
@@ -13,6 +12,7 @@ from .heuristics import DistanceHeuristic, Heuristic
 from .gracehandlers import ConcatenatingGraceHandler, GraceHandler
 from .jobhandlers import JobHandler, SerialJobHandler
 from .qevents import SilentQEvent
+from .qeventsequence import QEventSequence
 
 
 class QTarget:
